# app/api/matches.py
from fastapi import APIRouter, Query
from typing import List, Optional
from app.models.match import Match, WinProbability, WatchTarget, MatchPrediction, PredictionFactor

router = APIRouter(prefix="/matches", tags=["matches"])


@router.get("")
async def list_matches(
    league: Optional[str] = Query(None),
    status: Optional[str] = Query(None),
    date: Optional[str] = Query(None),
) -> dict:
    """List matches with optional filters by league, status, or date."""
    return {
        "success": True,
        "data": [],
        "error": None,
        "meta": {"total": 0}
    }


@router.get("/{match_id}")
async def get_match(match_id: str) -> dict:
    """Get match detail by ID."""
    return {
        "success": True,
        "data": {
            "id": match_id,
            "league": "T20",
            "team_a": "Team A",
            "team_b": "Team B",
            "status": "upcoming",
            "scheduled_at": "2026-06-25T19:00:00Z"
        },
        "error": None,
        "meta": {}
    }


@router.get("/{match_id}/prediction")
async def get_match_prediction(match_id: str) -> dict:
    """Get win probability + key factors + expected score for a match."""
    return {
        "success": True,
        "data": {
            "match_id": match_id,
            "win_probability": {
                "team_a_probability": 0.55,
                "team_b_probability": 0.45,
                "confidence": "medium",
                "key_factors": [
                    {
                        "factor": "Recent Form",
                        "impact": 0.3,
                        "description": "Team A won last 3 matches"
                    }
                ]
            },
            "expected_score": {
                "team_a": {"min": 160, "max": 180},
                "team_b": {"min": 155, "max": 175}
            }
        },
        "error": None,
        "meta": {}
    }


@router.get("/{match_id}/preview")
async def get_match_preview(match_id: str) -> dict:
    """Get AI-generated match preview using RAG/LLM."""
    return {
        "success": True,
        "data": {
            "match_id": match_id,
            "preview": "Preview content generated by RAG/LLM layer...",
            "analysis": "Detailed analysis with statistics...",
            "generated_at": "2026-06-24T16:00:00Z"
        },
        "error": None,
        "meta": {}
    }


@router.get("/{match_id}/watch")
async def get_watch_targets(match_id: str) -> dict:
    """Get legal watch targets for deep-linking to licensed providers."""
    return {
        "success": True,
        "data": {
            "match_id": match_id,
            "watch_targets": []  # Empty list is valid if no licensed target
        },
        "error": None,
        "meta": {}
    }